More progress on type checker, with small modifications to Syntax.
Added TNull type for the type checker to use when it envounters
a comment.

OpaqueType now derives Eq.

There is a lot of stuff I did for the type checker and this is
nowhere near done. I hope this is readable and you can deduce my
line of thinking.

One thing I noticed is that Funcs is reversed in the scheme
where the shader type comes first and the function after.
This might need to change to make things make a little more
sense.