Remove Haskell, Add Rust.
This commit removes all Haskell files and adds files important
for Cargo, as well as two files in Rust: main.rs, syntax.rs.

I have a basic working Syntax that I tried to make similar to
the previous abstract syntax with some improvements. I removed
the different types of comments and simply made it a vector of
strings. I also made Function definitions and Shader definiti-
ons part of the Expression syntax.

Notice too that the Block expression is simply a vector of
Boxed expressions. This is something I don't quite understand
yet but it is what I saw other people doing in their AST
implementations.

Please refer to this link for my source of inspiration:
https://createlang.rs/01_calculator/ast.html