Modify the Expr, Type. Remove Block, ShaderType, DiscreteType
Type now contains
	Int
	Uint
	Bool
	Float
	Double
	Array
	Struct
	Function

I don't know that we need a Tuple type, as that is just a 2
length array.

I added
	D
	Return
to Expr because we need a way to represent a double immediate
and I think that we can have a return statement. If we're
making this a stepping stone to GLSL then I think it could be
reasonable to allow that.