Add parse_binop, remove unneeded Box<>s in AST
parse_binop parses only the binary operator in a binary expression.
I did some trickery so that if we need to modify the representation
of the operator in the concrete syntax we only need to modify one
value in the function. If we want to change the number of operators
then that would require a little more tweaking, but not a significant
amount.

The simplification of the AST came from the relization that the size
of a Vec<T> is known at compile time, so we will not need to box either
the values inside Vec, nor the Vec itself. I changed the tests amd the
parsers to reflect this.